[WIP] Regression test for JIT crash
Regression test for the JIT crashing when compiling a method with
invalid dead dex code. For not compilable methods we don't gather
samples and we don't trigger JIT compilation. However
kAccDontBotherCompile is not persisted in the oat file and so we may end
up compiling a method which we shouldn't.

Test: m test-art-host-706-jit-skip-compilation
Bug: 31000839

Change-Id: Ib73de1888581bb7202474cfd7aca70af4cc2cc00