Update patch set 2
Patch Set 2:

Does the vdex file need to be treated separately from an oat file? Could you encapsulate the notion of vdex inside OatFile entirely, so that, even if vdex is a separate file on disk, logically it is written and loaded as part of writing and loading the oat file? That way, perhaps, you wouldn't have to expose vdex and vdex_filename in so many different places in the runtime, such as the OatFileAssistant.

Patch-set: 2