ART: Add verifier flags to oat file
If verifier detect the instruction which will always throw
the later instructions in the same basic blocks are considered as
dead code and are not verified while it may contains the "bad"
byte code. It is expected that such methods will never be compiled,
otherwise compiler may crash building the bad dead code.

If dex file was compiled by dex2oat and verification passed
at that moment that oat file will not contain the information
from verificaton affecting compilation of the method.
Specifically if method will throw at runtime and verifier
detects it then this information will be used by AOT compiler and
method will not be compiled. However at runtime,
verification of method will be ommitted because
verifier beleves to verification done at AOT moment and information
that method will always throw is omitted.
As a result if JIT works it will compile the method and may crash
at building stage.

To avoid this we need to pass verifiaction flags through oat file
to let JIT knows that no need to compiler such methods.

Change-Id: I85ebba660f56bd4425425f22769a3f224cef3fce
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>