Update patch set 2
Patch Set 2:

Unfortunately, no, I don't have any performance numbers to offer. All I can say is that the code for the 64-bit case is 4 instructions shorter, and one of the instructions removed is typically a very expensive operation (integer multiplication).

Are performance numbers going to be required going forward?

If so, is anyone doing the requisite benchmarks?

I know we have a few benchmarks for some of the other intrinsics which we've done internally. I don't know if they are ready to be (or can be) released externally.

Patch-set: 2
Reviewer: Gerrit User 1071873 <1071873@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0