Update patch set 1
Patch Set 1:

Hi Serguei,

Thanks for the CL. 

I also consider this approach but after discussions we decided against it. It does not worth adding the flags to the oat file.

I will upload a solution in which we will simply mark these methods as verified at runtime. That will fix the issue.

Patch-set: 1