[TEST] Benchmark DexCache string retrieval.
NOT INTENDED FOR SUBMISSION.
This benchmark relies on the dex cache string array using
a hash table of length 1024.

Benchmark the const-string when there is and when there
isn't a hash conflict in the DexCache string array.

Without
    https://android-review.googlesource.com/246591
the first execution of benchmarkConflict() takes about twice
as much time both on host and on Nexus 6 and the second
execution, with no memory to allocate, actually throws OOME.
No OOME is thrown with that CL.

Change-Id: I1ebeafda55a01efcbde1a9c88819b394af0fe485