Update patch set 1
Patch Set 1:

It's possible I'm misreading this code.  But I think this is assigning to the function argument, and has no lasting effect.  Presumably we intended this for the class member.

This line last changed in https://android-review.googlesource.com/#/c/111662.  But I believe C++ naming rules still had it changing the function argument prior to that, since the original checkin in https://googleplex-android-review.git.corp.google.com/#/c/254409.

If that analysis is correct, then I'm a bit nervous about this change, as presumably we've been happy with how things have been for the past four years.  Possibly instead of this CL, we should just remove the code within this constructor altogether?

Patch-set: 1