Update patch set 3
Patch Set 3:

I don't know about clang options to warn on this, but for what it's worth, master has 750+ cppcheck warnings of this sort.  A random spot check suggests many of them are methods which free a passed in pointer, and then set that pointer to NULL post-free.

Within the "art" silo, though, there only looks to be one other instance: art/patchoat/patchoat.cc:1093 where the bool base_delta_set is modified and never referenced again in the function.

Patch-set: 3