Don't age cards atomically.
Card aging of the entire table is non-atomic. Aging a card requires reading it,
looking at the value and if identical modifying it. After a card is read its
value may be changed by a mutator, such races are benign as the mutator's value
causes the collector to be more conservative for that card. If a card is already
marked then the GC may race with the mutator to age it, but as long as the age
information is age appropriate at the end of card aging then such races are
also benign.
Test: mm -j32 ART_TEST_GC_STRESS=true test-art-host

Change-Id: I49cf7490edde74d48913a90a9b83f52fd62ab932