Reduce number of physical pages used for DexCache's arrays
This commit reduces the amount of physical pages allocated for the .bss
section of an application's oat file.
Many of elements in the application's DexCache arrays are empty.
But during coping content of the DexCaches arrays from
the app's art file into the .bss section all pages become dirty.
The commit introduces a bitmap of DexCache arrays' pages. A bit is
set in the bitmap if the page of DexCache array contains at
least one non zero element. This bitmap is calculated at compile time
and stored into the app's art file. When the content of DexCache
arrays is copied from the app's art file into the bss section
we copy only pages with the bit set in the bitmap. Other pages
aren't touched thus the kernel does not allocates physical pages to
such content.
This approach allows to save 209Kb of memory after one minute of
GMaps use.
This number is the difference between the size of .bss section and
the Rss value from the /proc/<pid>/smaps file.

Change-Id: I3d7bed0805b95b5f344303581274ca8e8e69940e