Revert "Stop manually including gtest headers."
Broke all the tests, not 100% sure why yet. Building the
tests just hangs on my machine...

ccross was able to repro the failure and said it was a
double free issue. My guess is that there's some global
data in gtest that's being duplicated in libart-gtest.so
and in the test executable.

Reverting until I decide how I want to fix this. Shared
test libraries aren't really something we thought we
supported...

This reverts commit 728b27abf3c17cea3350a969d818b3d13d4604d4.

Change-Id: If932b96233b70627a65cfb9efb1cea8e03b95bd8