Remove custom CheckCast slow path caller saves for Baker CC.
For Baker CC, certain versions of CheckCast have both a
read-barrier marking slow path and a pCheckCast slow path.
Setting the custom slow path caller saves is perfectly valid
for the first but not for the second. Though it correctly
avoids reserving spill space that will not be used, it also
leads to marking live references in caller-save registers in
the register mask, causing a stack walk to try and retrieve
an unsaved register.

For the time being, revert to the default caller saves.

This is a partial revert of
    https://android-review.googlesource.com/254920

Test: Run ART test suite on host and Nexus 9.
Bug: 29231980
Bug: 30212852
Change-Id: I4e22125f3d8903c97506aa2e6e66bea8e8e6baef