Update patch set 1
Patch Set 1:

> Can you explain why you decided that empty should be considered
 > compressible?

I would say the all-ASCII string is a more common case than the non-ASCII, so that having the empty string as an all-ASCII will be helpful when allocating a new String from Strings. For example, new string is allocated from string A (empty) and B (ASCII). If empty string is treated as uncompressed, the new string object will be uncompressed which is kind of not true.

Patch-set: 1