Update patch set 2
Patch Set 2:

> Would it be manageable to add some debug check that would verify
 > the register mask somehow?

I'll have to think about that.

I think all the register saving avoidance based on CanThrowIntoCatchBlock() is dodgy and needs to be cleaned up. For DivZeroCheck and BoundsCheck I want to use the kSaveEverything and that will leave just the CheckCast in need of some special handling.

Patch-set: 2