ART: Use name-based hash for IMT layout
Use a simple name-based hash (based on full PrettyMethod)
for IMT layout. This means the slot will be stable, even if the
dex file layout changes.

The hash is obviously more computationally complex, which has an
effect on class-linking time. For boot image compilation on a
z840 in single-threaded mode, compilation increases by about half
a second (or about 4%):

Before: 17.3s
After:  17.9s

Bug:
Test: m
Test: m test-art-host
Test: device boots
Change-Id: I57ebf443ec0f93df84a02a370bc2e501c11fbb4a