Update patch set 1
Patch Set 1:

hmm ... might it be mma that doesn't pick it up correctly?

What I do is:
$mma test-art-target-gtest-...

I will try to figure out why if does fail for you but not for me (when it is supposed to fail). Anyway I believe that we need the exit 1 there (or false).

Patch-set: 1