Update patch set 1
Patch Set 1:

Unfortunately, ASAN is currently waiting on a fix before it boots correctly. It'd be good to have this in and ready before the fixes land though. I can run test-art-target on a non-ASAN target for this CL, to ensure that it doesn't break anything. Will amend the commit message once the test finishes running.

Patch-set: 1