Update patch set 1
Patch Set 1:

Agreed - the plan is actually to add TreeHugger boot tests for ASAN (it's been breaking because of changes in multiple different projects, so without the boot tests it's hard to keep track of how/where).  

For those tests we need both this fix, and a subsequent one that fixes build/Android.bp for b/29188876 - which will allow ASAN builds to complete (and thus allow for it to be tested).

Patch-set: 1