Update patch set 1
Patch Set 1:

No, the pointers in oat_filenames_ were becoming invalidated when the vector was resized, and then the invalid pointers were later used when trying to create the oat files, which resulted in either a failure to create the oat files, or garbage filenames.

I never saw this happen during an actual Android build - just when I was attempting to manually run dex2oat. I guess there was something about my environment that triggered this.

Patch-set: 1