Update patch set 2
Patch Set 2:

Sorry for the size of this CL.  Lots of tiny rewrites to use the new locations of the methods.  Thanks in advance!

David

P.S., Assuming the general feel of the change is acceptable, I'll spin the relevant tests out of utils_test.cc into other locations if desired.

Patch-set: 2
Label: Verified=0