Build: Avoid using ppid to generate temporary directory name.
When TMPDIR is left unset, use a munged version of
ANDROID_BUILD_TOP instead of PPID to construct a root directory
for all test runs. This workaround lets us run tests from multiple
repos at the same time, and will also avoid having to deal to
path length issues on the build bots.

This is a hacky workaround, but the only proper fix is stop using a
test runner built on top of make.

Bug: 32165422
Test: test-art-host

Change-Id: I61bc1eb3cf57d498f910573926d826a36020067a