Use the CLREX instruction in ARM and ARM64 CAS intrinsics.
Follow clang's implementation, which uses CLREX in
compare-and-exchange operations on the failure path, i.e.
when the value read by the LDREX (ARM) or LDAXR (ARM64)
instruction is not the expected value, in order to release
the monitor.  The previous implementation was perfectly
correct, but this one should may improve performance on some
micro-architectures.

Test: make test-art-target-run-test-004-UnsafeTest
Change-Id: I1138156ca5a615739dd6a96411db719f95ce67c1