Save the non-verified classes in the VerifierDeps.
We will need that information when taking an OTA to make sure
the same set of classes need to be verified at runtime.

Currently, the vdex file wikk contain a list of unverified
classes. We could alternatively encode a bit vector of the size
of the type_id array, but the few experiments I did show that
the bit vector is actually larger. We can refine this later.

bug: 30937355
test: m test-art-host
test: verifier_deps_test.cc

Change-Id: I2670e4fd2e54ee7a148246baa705fda3a56617ff