Update patch set 2
Patch Set 2:

> > I could push the IsInBlock() check into the iterator used by BCE.
 > > This works, since the next field is not changed on replacement.
 > Or
 > > I could do a callback when each instruction is replaced and
 > adjust
 > > the current (no need to specifically do that on the next in that
 > > case). Before coding this out, please let me know your
 > preference.
 > 
 > Hmmm, thinking about it more, we have a deeper problem that the
 > next instruction of the dead HBoundsCheck will be null, therefore
 > skipping the entire next instructions in the block.
 > 
 > Maybe we should buffer the list of dead HBoundsCheck and remove
 > them just at the end?

No the next pointer is not reset on deletion (DavidB and I talked about this a while back whether we could keep on relying on that). But rather than buffering, which seems ugly, I would prefer the second proposal I made to keep the iterator valid. I already use a single method for all replacement/deletion so that should be least intrusive.....

Patch-set: 2