WIP: inlining a function that does not return
So far: a regression test that reproduces the failure discovered by the
scanner.  This illustrates that the issue with LinearOrder failing is
due to incorrect loop information being computed for blocks that are
not on a path to the inlined function's return.  They should not be
considered part of the caller's enclosing loop, but are today.

Bug: 32547653
Test: run-test --host 478-checker-inline-noreturn
Change-Id: I9694a1cb861430051c801d07f7ce29752332cba5