Update patch set 1
Patch Set 1:

I think one the peculiarities of the ARM64 back end is that D31 was (before this CL) used both as a scratch register by the parallel move resolver (see ParallelMoveResolverARM64::AllocateScratchLocationFor) *and* VIXL's FP scratch register pool, solicited by CodeGeneratorARM64::MoveLocation, in the case of move between stack slots to provide a temporary register.

FYI, the problematic parallel move that was encountered is:

      0    0    v5669  ParallelMove liveness:3856 moves:[d0->124(sp),140(sp)->144(sp),124(sp)->128(sp),128(sp)->132(sp),d4->d5,d5->d6,d6->d7,d7->d16,132(sp)->136(sp),136(sp)->140(sp),d18->d19,d19->d20,144(sp)->d21,d21->d28,d28->d29,d29->d30,d30->d0,d2->d1,d16->d4,d1->d18,d20->d2] liveness:3856 loop:none

The assertion is failing because the second move (`140(sp)->144(sp)`) is a move between stack slots that was requiring D31, which is already used to break a cycle, if I understand ParallelMoveResolverNoSwap::PerformMove correctly.

I'll try to come up with a reduce test case and implement it as a run-test.

Patch-set: 1
Label: Verified=0