Update patch set 1
Patch Set 1:

Answering my own questions:

> Why isn't this a problem for the partial/full case?

As mentioned, this is okay because if the object is reachable, the object would be on the mark stack and its class will be marked.

> Why isn't this a problem for other non-allocation operations like instanceof or a static call?

Non-allocation operations like instanceof or a static call have to be done reflectively from outside of the class loader boundary or non-reflectively within the class loader boundary. In both cases, if those operations are taking place, GC won't unload the class loader because the class used by the operations should be reachable.

Patch-set: 1
Label: Verified=0