[WIP] Make sure that const-class linkage is preserved.
Make sure that a const-class instruction always returns the
same Class<?> as it returned the first time, i.e. when it
was "linked", even for class loaders that are not
well-behaved, even if the dex cache type array is cleared.
This is in preparation for changing to hash-based array.

TODO: Add loaded classes to ClassTable when needed.
      (The test is currently failing.)

Test: 626-const-class-linking
Change-Id: Ie9b721464b4e9a5dcce8df8095548e983bba1fe8