Update patch set 1
Patch Set 1:

> > Has something in the runtime changed recently to require a change
 > > in how this test is written? What problem are you trying to solve
 > > with this CL?
 > 
 > We had one failure, see dalvik-team@ emails where Andreas diagnosed
 > it:
 > 
 > "039 is a not-sufficiently synchronized test. But it seems to point
 > to a pretty badly loaded build bot (or a really bad scheduler),
 > which could conceivably lead to a race in the 030 test, too.
 > 
 > (BTW, 030 is bad wrt/ local variables. dx/jack could keep the
 > register alive...)"
 > 
 > I cannot access the logs right now due to some outage.

@Richard:

The failure was that the UNREACHABLE code was executed:

 #################### diffs
 --- expected.txt	2016-11-09 00:49:01.284124596 +0000
 +++ output.txt	2016-11-09 01:39:42.839428700 +0000
 @@ -2,3 +2,5 @@
  Finalizer started and spinning...
  Finalizer done spinning.
  Finalizer sleeping forever now.
 +E/System  (27243): Main$BadFinalizer.finalize() timed out after 10 seconds
 +UNREACHABLE
 ####################

My best attempt of understanding this was that the main thread was scheduled a lot more often than the finalizer and watchdog thread. As the main thread isn't waiting "infinitely," it is conceivable that the race is:

  1) Main thread triggers finalization
  2) Main thread runs through all GCs and its sleep
  3) The finalizer watchdog wakes up and prints the error line
  4) The main thread progresses to the unreachable code

The problem is that we don't have scheduling guarantees (we could increase the watchdog priority to make it less likely, but I don't think that's a great idea).

So I thought it's better to just trigger the GCs, make sure that the finalizer has started running, and then go to sleep "infinitely" (modulo a watchdog).

Patch-set: 1