WIP Elide read barriers for HClinitCheck/HLoadClass.
It's safe for an HClinitCheck to read a class status from a class
loaded by a HLoadClass without a read barrier. This is because even if
we redundantly take the HClinitCheck slow path due to a stale class
status value, the slow path will reload the class with a read barrier
and see the up-to-date class status.

Bug: 29516974
Bug: 12687968
Test: TBD
Change-Id: I37e2f51f7757022bbed27c5d25412b2f891b4ec2