Flush updated JIT code into shared memory before invalidating instruction pipelines on active cores with mprotect(RWX->RW)
According to the ARM documentation
(http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.den0024a/BABJDBHI.html)
cleaning a cache or cache line means writing the contents of cache
lines that are marked as dirty out to the next level of the cache,
or to the main memory.
So we need to flush the caches before removing write permission for jit-code-cache
memory region.

Test: run test-art-host

Change-Id: I70cbb0dc2ea37e42f2011d23333e386ae3f639c7