Fix fault handler to distinguish native and oat methods better
Basically the x86 fault handler thinks that if the top of stack
contains a pointer that looks like an ArtMethod* (i.e. aligned and
contains a valid class pointer) then it is treated as such and
used to find its OatMethodHeader and translate native PC to dex_pc.
But there could be cases when the stack and native PC accidentally
satisfy the safe checks (alignmet and class verification) but still
correspond to a native method. In these cases behavior of ART is
undefined and further crash reports may contain misleading
messages.

The patch adds a safe check that the PC stays inside the address
range of the identified OatMethod method. In addition the PC check
in the stack walk routine is substituted with the safe
OatMethodHeader detection to not duplicate the code.

Test: m test-art-host
Change-Id: I7282f917f31f56cc9c611e9b75dd626229b69da4
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>