Update patch set 4
Patch Set 4:

(1 comment)

I spoke with Andreas, and he noted that we're missing a test for this new filter being used in dex2oat. I'll add one, though probably in the next CL or I'd have to write the test and immediately change its expected results in the next CL in this series.

Patch-set: 4
Label: Verified=0