Verify for type_id out of range errors
The dalvik documentation states that several fields used as type ids are
uint, which is 32 bits.  Other places use uint16, which is obviously 16.
This adds assertions that there are fewer than 65536
class_def_items/class_data_items and that the fields in question are
sign-extended to 32 bits from a 16 bit value.  (I had to checkf for sign
extension, because it seems zero-extension fails...)

Test: m test-art-host
Change-Id: I584615192086e6d8bb30aa11217d569d6dbe1280