Update patch set 4
Patch Set 4: Code-Review+1

Sorry for replying late to this. But this looks like a correct fix.

AFAIK mprotect() will remove some of the attributes from a TLB walk. Thus if a later cache invalidation or another memory operation  that writes to the range covered by a mprotect() we might hit the issue described here.

I should take a look at how mprotect() works to validate this. I will have a chat with our Kernel guys and come back with you if this is not the case.

Thanks for fixing this!

Patch-set: 4